# api/swagger.yaml
openapi: 3.0.0
info:
  title: DevShop API - OpenAPI 3.0
  description: |-
    This is a test API that simulates an ecommerce API interaction, made for learning how to use Node.js and it's technologies ecosystem for the Backend. hope it helps.

    For completing request to protected endpoints, you need to authenticate, click the Authorize green button and give it this values:
      - **JWTToken  (http, Bearer)**: Use a new token logging in ``/auth/login`` or use this token for the meantime:

        <details>
        <summary>Show the sample JWT</summary>
            ``eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjUsInJvbGUiOiJhZG1pbiIsImlhdCI6MTY5NzkxOTU0MH0.jZcLU3oMUGAKordYnhDdMo8eL-8Jh5XfeJtgseICSug``
        </details>

    Some useful links:
    - [The DevShop repository](https://github.com/Lenoxo/DevShop)
    - [The source API definition for the DevShop](pending...)
  contact:
    name: Emanuel Padilla (Lenoxo)
    url: https://emanueldev.vercel.app/
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  version: 1.0.0
servers:
  - url: http://localhost:{port}/api/v1
    description: Local server direction.
    variables:
      port:
        default: "3000"
  - url: https://backend-ecommerce-production-a555.up.railway.app/api/v1
    description: Production server url.
paths:
...
  /customers:
    get:
      summary: Get all customers
      operationId: getCustomers
      tags:
        - Customers
      responses:
        "200":
          description: Successful retrieval of customers
          content:
            application/json:
              schema:
                type: array
                example:
                  - id: 1
                    name: emanu
                    lastName: pelaez
                    phone: "123123154"
                    createdAt: "2023-09-23T03:32:16.170Z"
                    userId: 5
                    user:
                      id: 5
                      email: emanu1@example.com
                      password: $2b$10$mMfF1HGJX/uBu3JAUw.iHuVlasq4OJll8LR.0UFTaE4eRICa8/Tsi
                      role: admin
                      recoveryToken: null
                      createdAt: "2023-09-23T03:32:16.170Z"
                  - id: 3
                    name: harry
                    lastName: olimpus
                    phone: "15154595792"
                    createdAt: "2023-09-24T19:43:12.159Z"
                    userId: 10
                    user:
                      id: 10
                      email: harry@example.com
                      password: $2b$10$U4dauOa8o.Hjoq.8ls9NMebHdE.3No4y7LmmmHO9N6d17LtmwD5Uq
                      role: customer
                      recoveryToken: null
                      createdAt: "2023-09-24T19:43:12.159Z"
    post:
      summary: Create a new customer
      operationId: createCustomer
      tags:
        - Customers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                name: test
                lastName: ramirez
                phone: "54545454523"
                user:
                  name: test
                  email: thisisnotadrill@example.com
                  password: example again 23
                  role: seller
      responses:
        "201":
          description: Successful creation of a new customer
          content:
            application/json:
              schema:
                type: object
                example:
                  createdAt: "2023-10-21T23:14:06.091Z"
                  id: 4
                  name: test
                  lastName: ramirez
                  phone: "54545454523"
                  user:
                    createdAt: "2023-10-21T23:14:06.091Z"
                    id: 14
                    email: thisisnotadrill@example.com
                    role: seller
                    recoveryToken: null
                  userId: 14
  /customers/{id}:
    get:
      summary: Get a customer by ID
      operationId: getCustomerById
      tags:
        - Customers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful retrieval of the customer
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 1
                  name: emanu
                  lastName: pelaez
                  phone: "123123154"
                  createdAt: "2023-09-23T03:32:16.170Z"
                  userId: 5
                  user:
                    id: 5
                    email: emanu1@example.com
                    password: $2b$10$mMfF1HGJX/uBu3JAUw.iHuVlasq4OJll8LR.0UFTaE4eRICa8/Tsi
                    role: seller
                    recoveryToken: null
                    createdAt: "2023-09-23T03:32:16.170Z"

    patch:
      summary: Update a customer by ID
      operationId: updateCustomer
      tags:
        - Customers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                lastName: mendez
      responses:
        "200":
          description: Successful update of the customer
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 1
                  name: emanu
                  lastName: mendez
                  phone: "123123154"
                  createdAt: "2023-09-23T03:32:16.170Z"
                  userId: 5
                  user:
                    id: 5
                    email: emanu1@example.com
                    password: $2b$10$mMfF1HGJX/uBu3JAUw.iHuVlasq4OJll8LR.0UFTaE4eRICa8/Tsi
                    role: seller
                    recoveryToken: null
                    createdAt: "2023-09-23T03:32:16.170Z"
    delete:
      summary: Delete a customer by ID
      operationId: deleteCustomer
      tags:
        - Customers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful deletion of the customer
          content:
            application/json:
              schema:
                type: object
                example:
                  id: 5

  # Orders paths
  /orders:
    get:
      summary: Get all orders
      operationId: getAllOrders
      tags:
        - Orders
      responses:
        "200":
          description: Successful retrieval of orders
          content:
            application/json:
              schema:
                type: array
                example:
                  - totalPrice: 0
                    id: 1
                    customerId: 3
                    date: "2023-09-25T19:13:24.145Z"
                    customer:
                      id: 3
                      name: harry
                      lastName: olimpus
                      phone: "15154595792"
                      createdAt: "2023-09-24T19:43:12.159Z"
                      userId: 10

    post:
      summary: Create a new order
      operationId: createOrder
      security:
        - JWTToken: []
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                customerId: 1
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                example:
                  totalPrice: 0
                  date: "2023-10-21T23:59:25.844Z"
                  id: 4
                  customerId: 1
  /orders/{id}:
    get:
      summary: Get a specific order by ID
      operationId: getOrderById
      tags:
        - Orders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful retrieval of the order
          content:
            application/json:
              schema:
                type: object
                example:
                  totalPrice: 66
                  id: 1
                  customerId: 3
                  date: "2023-09-25T19:13:24.145Z"
                  customer:
                    id: 3
                    name: harry
                    lastName: olimpus
                    phone: "15154595792"
                    createdAt: "2023-09-24T19:43:12.159Z"
                    userId: 10
                    user:
                      id: 10
                      email: harry@example.com
                      password: $2b$10$U4dauOa8o.Hjoq.8ls9NMebHdE.3No4y7LmmmHO9N6d17LtmwD5Uq
                      role: customer
                      recoveryToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjEwLCJpYXQiOjE2OTU2ODg1NjUsImV4cCI6MTY5NTY4OTQ2NX0.kuiFUPHlMQCuDQI1llE0kqbNasezh-qwjgI2CvbzsZ4"
                      createdAt: "2023-09-24T19:43:12.159Z"
                  items:
                    - id: 3
                      name: zapatos de jazz
                      description: "Esta es una prueba de producto, no importa su contenido."
                      price: 11
                      image: https://image.com/example
                      categoryId: 6
                      createdAt: "2023-09-25T19:14:28.681Z"
                      OrderProduct:
                        id: 1
                        orderId: 1
                        productId: 3
                        amount: 5
                        date: "2023-09-25T19:35:14.323Z"
  /orders/add-item:
    post:
      summary: Add an item to an order
      operationId: addItemToOrder
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                productId: 3
                orderId: 2
                amount: 3
      responses:
        "201":
          description: Item added to the order
          content:
            application/json:
              schema:
                type: object
                example:
                  date: "2023-10-22T00:12:11.790Z"
                  id: 3
                  productId: 3
                  orderId: 2
                  amount: 3

  # Products paths
...
components:
...
  securitySchemes:
    JWTToken:
      type: http
      scheme: bearer
      bearerFormat: JWT